loan_management2/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
distlib==0.3.7
Django==4.2.7
djangorestframework==3.14.0
filelock==3.13.1
platformdirs==4.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.24.7
